# pkg/response/response.go
package response

import (
	"gin-ddd-example/pkg/errors/e"
	"net/http"

	"github.com/gin-gonic/gin"
)

// Response 代表一个响应给客户端的消息结构，包括错误码，错误消息，响应数据
type Response struct {
	Code    int         `json:"code"`           // 错误码，0表示无错误
	Message string      `json:"message"`        // 提示信息
	Data    interface{} `json:"data,omitempty"` // 响应数据，一般从这里前端从这个里面取出数据展示
}

// JSON 发送json格式的数据
func JSON(c *gin.Context, errCode int, data interface{}) {
	// 如果code != 0, 失败的话 返回http状态码400（一般也可以全部返回200）
	// 返回400 更严谨一些，个人接触的项目中大部分都是400。
	var httpStatus int
	if errCode != e.Success {
		httpStatus = http.StatusBadRequest
	} else {
		httpStatus = http.StatusOK
	}

	c.JSON(httpStatus, Response{
		Code:    errCode,
		Message: e.GetMsg(errCode),
		Data:    data,
	})
}